% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exInput}
\alias{exInput}
\title{1500 random lines from the FeederWatch interaction database.}
\format{A data frame with 1500 rows and 2 variables:
\describe{
  \item{source}{the identity of the initiating animal}
  \item{target}{the identity of the receiving animal}
  ...
}}
\source{
\url{http://feederwatch.org/}
}
\usage{
exInput
}
\description{
A dataset containing random lines of FeederWatch interaction datas.
}
\keyword{datasets}
