% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepAON.R
\name{prepAON}
\alias{prepAON}
\title{Prep data for AON plot}
\usage{
prepAON(ranks, edge.list)
}
\arguments{
\item{ranks}{Numeric vector with names matching those in edge.list. No check is
made to ensure these match 1:1, onus on user.}

\item{edge.list}{A data frame of observations, with the first column as the dominant
species/individual and the second column as the subordinate species/individual.}
}
\value{
A list object, prepped for use in plotting an attribute-ordered network.
}
\description{
Prepare data for an attribute-ordered network plot.
}
\details{
User should format edge.list so it looks like the example data (exInput).
}
\examples{
#load in the example data
data(exInput)

#run the Bradley Terry function. takes a few seconds to run.
results <- myBTM(edge.list=exInput, cores=2)

#the names of these scores have weird periods in them. use agrep to solve
species <- unique(c(as.character(exInput$source), as.character(exInput$target)))
names(results) <- sub("..", "", x=names(results))

for(i in 1:length(results))
{
  names(results)[i] <- species[agrep(names(results)[i], species, max.distance=0.1)]
}

prepped <- prepAON(results, exInput)
}
\references{
Hobson, E. A., D. J. John, T. L. Mcintosh, M. L. Avery, and T. F. Wright.
2015. The effect of social context and social scale on the perception of
relationships in monk parakeets. Current Zoology 61:55-69.
}
\author{
Elizabeth A. Hobson, https://sites.google.com/site/hobsonresearch/
}
