% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transCruncher.R
\name{transCruncher}
\alias{transCruncher}
\title{Summarize results of transitivityChecker}
\usage{
transCruncher(network.size, conservative, read.wd)
}
\arguments{
\item{network.size}{The size of the subnetworks over which the user would like to
summarize network transitivity.}

\item{conservative}{Whether to summarize results over calls to transitivityChecker
that were thresholded according to a conservative (TRUE) or a liberal (FALSE) rule.}

\item{read.wd}{The path where the results are saved.}
}
\value{
A list of result objects. First, a list, where each element in that list is
a data frame of results at a given threshold. That data frame provides the number of
total subnetworks that each species was involved in, how many of those were
intransitive, and the proportion of intransitivie subnetworks each was involved in.
Second, a data frame summarizing the total number of transitive, intransitive, and
NA subnetworks in the larger network.
}
\description{
Use previously run results of transitivityChecker to summarize transitivity of
networks, possibly over multiple runs of transitivityChecker.
}
\details{
This function is used to summarize the transitivity of a larger network as
assessed at a given subnetwork size and tie-handling method (conservative or not).
For example, a series of analyses at the same subnetwork size and tie-handling method
but varying threshold might be run, then this function would summarize the results.
}
\examples{
\dontrun{
#load in the example data
data(exInput)

#look at transitivity of dyads
transitivityChecker(disp.input=exInput, network.size=2, cutoff=0.5,
  conservative=TRUE, write.wd=tempdir(), cores=4)

#assess transitivity of these dyads
transCruncher(2, TRUE, tempdir())
}
}
\references{
Miller, E. T., D. N. Bonter, C. Eldermire, B. G. Freeman, E. I. Greig,
L. J. Harmon, C. Lisle, and W. M. Hochachka. 2017. Fighting over food unites the
birds of North America in a continental dominance hierarchy.
biorxiv https://doi.org/10.1101/104133
}
