% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitivityChecker.R
\name{transitivityChecker}
\alias{transitivityChecker}
\title{Run the DAG method over a social network}
\usage{
transitivityChecker(disp.input, network.size, cutoff, conservative, write.wd,
  cores)
}
\arguments{
\item{disp.input}{An edge-list-like data frame, formatted like the exInput data.}

\item{network.size}{The subnetwork size (dyad, triad, quartet, etc.) at which to
assess transitivity.}

\item{cutoff}{The proportion of interactions a species needs to have won in order to
be considered the winner.}

\item{conservative}{TRUE or FALSE. Whether a tied interaction is set to a
bidirectional edge or the edges are removed entirely.}

\item{write.wd}{The path to the working directory where results will be written.}

\item{cores}{The number of cores to employ for parallel processing. Set to 'seq' to
run sequentially (i.e. not in parallel).}
}
\value{
Nothing to the workspace. Writes two RDS files to write.wd.
}
\description{
Calculate transitivity of a network using the directed acyclic graph method.
}
\details{
Takes a larger dominance hierarchy/social network, and splits it into all
possible subnetworks of the specified size. Simplifies each of those networks
according to the arguments provided, then assesses subnetwork transitivity. Writes
all results to write.wd. These results can be summarized in subsequent steps with
transCruncher().
}
\examples{
#load in the example data
\dontrun{
data(exInput)

#look at transitivity of dyads
transitivityChecker(disp.input=exInput, network.size=2, cutoff=0.5,
  conservative=TRUE, write.wd=tempdir(), cores=4)

#here are all the possible subnetworks
allPoss <- readRDS(paste(tempdir(), "combn_results_2species_cutoff0.5.RDS",
  sep="/"))

#and here are transitivity results
transResults <- readRDS(paste(tempdir(),
   "trans_results_2species_cutoff0.5consTRUE.RDS", sep="/"))

#here are all transitive networks we had sufficient information to assess
allPoss[,which(transResults==TRUE)]
}
}
\references{
Miller, E. T., D. N. Bonter, C. Eldermire, B. G. Freeman, E. I. Greig,
L. J. Harmon, C. Lisle, and W. M. Hochachka. 2017. Fighting over food unites the
birds of North America in a continental dominance hierarchy.
biorxiv https://doi.org/10.1101/104133
}
