% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transPlotter.R
\name{transPlotter}
\alias{transPlotter}
\title{Plot function for transitivity results}
\usage{
transPlotter(results.list, color)
}
\arguments{
\item{results.list}{A list of results from multiple calls to transitivityChecker.}

\item{color}{Whether to plot each result in a different color, or a different
line type.}
}
\value{
An R plot.
}
\description{
Plot the summarized results from multiple calls to transitivityChecker.
}
\details{
Uses a Brewer color palette to plot the results of multiple calls to
transitivityChecker(). Each unique subnetwork size gets its own color.
}
\examples{
#load in the example data
\dontrun{
data(exInput)

#look at transitivity of dyads. this whole thing will take a few minutes
#to run. be patient.
transitivityChecker(disp.input=exInput, network.size=2, cutoff=0.5,
  conservative=TRUE, write.wd=tempdir(), cores=4)

transitivityChecker(disp.input=exInput, network.size=2, cutoff=0.6,
  conservative=TRUE, write.wd=tempdir(), cores=4)

transitivityChecker(disp.input=exInput, network.size=2, cutoff=0.7,
  conservative=TRUE, write.wd=tempdir(), cores=4)

transitivityChecker(disp.input=exInput, network.size=2, cutoff=0.8,
  conservative=TRUE, write.wd=tempdir(), cores=4)

transitivityChecker(disp.input=exInput, network.size=2, cutoff=0.9,
  conservative=TRUE, write.wd=tempdir(), cores=4)

transitivityChecker(disp.input=exInput, network.size=2, cutoff=1.0,
  conservative=TRUE, write.wd=tempdir(), cores=4)

#now look at transitivity of a couple triad thresholds
transitivityChecker(disp.input=exInput, network.size=3, cutoff=0.5,
  conservative=TRUE, write.wd=tempdir(), cores=4)

transitivityChecker(disp.input=exInput, network.size=3, cutoff=1.0,
  conservative=TRUE, write.wd=tempdir(), cores=4)

twoSpp <- transCruncher(2, TRUE, tempdir())
threeSpp <- transCruncher(3, TRUE, tempdir())

#bind together just the count results for plotting
toPlot <- list(twoSpp$counts, threeSpp$counts)

transPlotter(toPlot, color=TRUE)
}
}
\references{
Miller, E. T., D. N. Bonter, C. Eldermire, B. G. Freeman, E. I. Greig,
L. J. Harmon, C. Lisle, and W. M. Hochachka. 2017. Fighting over food unites the
birds of North America in a continental dominance hierarchy.
biorxiv https://doi.org/10.1101/104133
}
